requirements.txt
# Streamlit Dashboard Requirements
# For deployment to Streamlit Cloud

# Core Dashboard
streamlit==1.40.1
altair>=5.0.0
python-dotenv>=1.0.0
requests>=2.28.0  # For downloading data from external storage

# Data Processing
pandas>=2.0.0
numpy>=1.20.0
pyarrow>=10.0.0  # Required for reading Parquet files

# NLP
nltk>=3.8.1

# OpenAI for RAG+LLM (Legacy 0.28.x compatibility)
openai==0.28.0

# Optional: for TF-IDF retrieval method
scikit-learn>=1.3.0

